+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v1.20.0...HEAD)
 ------------------------------------------------------------------------
 ### Security
+- Runs `npm audit fix`
 - Removes `webpack-pwa-manifest` from `devDependencies`
 
 [1.20.0](https://github.com/jshwi/jss/releases/tag/v1.20.0) - 2022-07-01