+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/jss/compare/v1.20.4...HEAD)
+[Unreleased](https://github.com/jshwi/jss/compare/v1.20.5...HEAD)
+------------------------------------------------------------------------
+
+[1.20.5](https://github.com/jshwi/jss/releases/tag/v1.20.5) - 2022-11-16
 ------------------------------------------------------------------------
 ### Security
 - Bumps `loader-utils` from 2.0.3 to 2.0.4